+++ src/index.js
@@ -1,9 +1,0 @@
-// @flow strict
-// eslint-disable-next-line filenames/match-exported
-import CookieApi from './CookieApi';
-
-require('cross-fetch/polyfill');
-
-export default CookieApi;
-
-export { CookieApi };