+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [2.0.0] 2018-11-16
+### Changed
+- Convert repo from Flow to Typescript
+- Change to default import
+
+
 ## [1.1.0] 2018-06-11
 ### Added
 - Added ESdocs config